Do not debounce changes to number inputs
For some fields (text in particular) we only send change to
the server on blur. In case of type="number" inputs, clicking
the up/down arrows does not focus the input in Firefox, which
means the input will not receive blur either. We decided to
not debounce changes to number inputs, since those changes
are not as intense as text.